+++ CS2 Metro/VersionInfo.cs
@@ -2,6 +2,6 @@ namespace CS2_Metro
 {
     internal static class VersionInfo
     {
-        public const string ReleaseVersion = "v0.1.0-alpha.3";
+        public const string ReleaseVersion = "v0.1.0-alpha.7";
     }
 }